+++ sdk/identity/identity/src/msal/nodeFlows/msalAuthorizationCode.ts
@@ -32,32 +32,24 @@ export class MsalAuthorizationCode extends MsalNode {
     if (options.clientSecret) {
       this.msalConfig.auth.clientSecret = options.clientSecret;
     }
-    console.log("ZZZZZZ", this.msalConfig.auth.clientSecret);
   }
 
-  async getAuthCodeUrl(options: { scopes: string[], redirectUri: string }): Promise<string> {
+  async getAuthCodeUrl(options: { scopes: string[]; redirectUri: string }): Promise<string> {
     await this.init();
     return this.confidentialApp!.getAuthCodeUrl(options);
   }
 
   protected async doGetToken(scopes: string[], options?: GetTokenOptions): Promise<AccessToken> {
     try {
-      console.log("BBBBB", {
-        scopes,
-        redirectUri: this.redirectUri,
-        code: this.authorizationCode
-      })
       const result = await this.confidentialApp?.acquireTokenByCode({
         scopes,
         redirectUri: this.redirectUri,
         code: this.authorizationCode
-      }); 
-      console.log({ result });
+      });
       // The Client Credential flow does not return an account,
       // so each time getToken gets called, we will have to acquire a new token through the service.
       return this.handleResult(scopes, this.clientId, result || undefined);
     } catch (err) {
-      console.log("MSAL ERROR", err);
       throw this.handleError(scopes, err, options);
     }
   }